[refactor]デスティネーションのルートに関するsealedクラスを作成
