+++ app/src/main/java/com/example/pokebook/ui/screen/BottomNavigationScreen.kt
@@ -30,27 +30,6 @@ import com.example.pokebook.ui.viewModel.SearchViewModel
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 /**
- * ボトムナビゲーション
- */
-sealed class BottomNavItems(
-    val route: String,
-    val name: String,
-    val icon: ImageVector
-) {
-    object Home : BottomNavItems("home", "HOME", Icons.Filled.Home)
-    object Search : BottomNavItems("search", "SEARCH", Icons.Filled.Search)
-    object Like : BottomNavItems("like", "LIKE", Icons.Filled.Star)
-    object Setting : BottomNavItems("setting", "SETTING", Icons.Filled.Settings)
-}
-
-val navItems = listOf(
-    BottomNavItems.Home,
-    BottomNavItems.Search,
-    BottomNavItems.Like,
-    BottomNavItems.Setting
-)
-
-/**
  * NavHost に宛先を設定する
  */
 @SuppressLint("ComposableDestinationInComposeScope")
