+++ app/src/main/java/com/example/pokebook/ui/screen/ScreenRoutes.kt
@@ -1,0 +1,28 @@
+package com.example.pokebook.ui.screen
+
+/**
+ * HOMEタブ
+ */
+sealed class HomeScreen(val route: String) {
+    object PokemonListScreen : HomeScreen("home/pokemonListScreen")
+    object PokemonDetailScreen : HomeScreen("home/pokemonDetailScreen")
+}
+
+/**
+ * SEARCHタブ
+ */
+sealed class SearchScreen(val route: String) {
+    object SearchTopScreen : SearchScreen("search/searchScreen")
+    object PokemonListScreen : SearchScreen("search/pokemonListScreen")
+    object PokemonNotFound : SearchScreen("search/pokemonNotFound")
+    object PokemonDetailScreen : SearchScreen("search/pokemonDetailScreen")
+}
+
+/**
+ * LIKEタブ
+ */
+
+
+/**
+ * SETTINGタブ
+ */
